requirements.txt
Flask==3.1.1
requests==2.32.4
python-dotenv==1.1.1
numpy==2.3.2
scikit-learn==1.7.1